+++ src/sub-clients/adminLinks/model/AdminLinks.ts
@@ -2,6 +2,7 @@ import { LinkEntity } from '../../common/model/LinkEntity';
 
 export interface AdminLinks {
   collection: LinkEntity;
+  video: LinkEntity;
   createCollection: LinkEntity;
   adminCollectionSearch: LinkEntity;
   jobs: LinkEntity;