+++ src/sub-clients/videoTypes/model/VideoType.ts
@@ -1,0 +1,4 @@
+export interface VideoType {
+  id: number;
+  name: string;
+}