+++ src/sub-clients/videos/client/VideosClient.ts
@@ -1,0 +1,8 @@
+import { VideoWithBoclipsProjection, Video } from '../model/Video';
+
+export interface VideosClient {
+  get(id: string): Promise<Video>;
+  hasBoclipsProjection(
+    video: VideoWithBoclipsProjection | Video,
+  ): video is VideoWithBoclipsProjection;
+}