+++ src/sub-clients/videos/utils/hasBoclipsProjection.ts
@@ -1,0 +1,6 @@
+import { VideoWithBoclipsProjection, Video } from '../model/Video';
+
+export const hasBoclipsProjection = (
+  video: VideoWithBoclipsProjection | Video,
+): video is VideoWithBoclipsProjection =>
+  (video as VideoWithBoclipsProjection).contentPartner !== undefined;