+++ src/test-support/VideosFactory.ts
@@ -1,4 +1,7 @@
-import { Video } from '../sub-clients/common/model/Video';
+import {
+  VideoWithBoclipsProjection,
+  Video,
+} from '../sub-clients/videos/model/Video';
 
 export class VideoFactory {
   public static sample = (video: Partial<Video>): Video => ({
@@ -6,16 +9,43 @@ export class VideoFactory {
     ...video,
   });
 
-  private static defaultFields = (): Video => ({
+  protected static defaultFields = (): Video => ({
     id: 'video-123',
     title: 'video title',
     description: 'video description',
-    contentPartner: 'some content partner',
-    disabledDistributionMethods: [],
+    releasedOn: new Date(),
     playback: { type: 'STREAM' },
+    subjects: [],
+    badges: [],
+    legalRestrictions: '',
+    ageRange: {},
+    rating: 4,
+    yourRating: undefined,
+    bestFor: undefined,
+    createdBy: undefined,
+    promoted: false,
+    language: undefined,
     links: {
+      self: null,
+      logInteraction: null,
       update: null,
     },
-    subjects: [],
+  });
+}
+
+export class VideoWithBoclipsProjectionFactory extends VideoFactory {
+  public static sample = (
+    video: Partial<VideoWithBoclipsProjection>,
+  ): VideoWithBoclipsProjection => ({
+    ...VideoWithBoclipsProjectionFactory.defaultFields(),
+    ...video,
+  });
+
+  protected static defaultFields = (): VideoWithBoclipsProjection => ({
+    ...VideoFactory.defaultFields(),
+    contentPartner: 'some content partner',
+    contentPartnerId: 'partner-1',
+    contentPartnerVideoId: 'video-id',
+    type: { id: 1, name: 'INSTRUCTIONAL' },
   });
 }